[MLIR][OpenMP] Add omp.loopnest operation
This patch introduces the `omp.loopnest` MLIR operation, which contains
the information of a rectangular collapsed loop nest. It mirrors the existing
representation of `omp.wsloop`, `omp.simdloop` and `omp.taskloop`, with the
intent of removing loop information from these in a subsequent patch.

This representation is a temporary solution that does not address loop
transformations. That is the goal of the `omp.canonical_loop` discussion.