[MLIR][OpenMP] Add simplified omp.canonical_loop
This patch introduces the `omp.canonical_loop` MLIR operation, which contains
the information of a collapsed loop nest. It mirrors the existing
representation of `omp.wsloop`, `omp.simdloop` and `omp.taskloop`, with the
intent of removing loop information from these in a subsequent patch.

This representation is a temporary solution that does not address loop
transformations.