[MLIR][OpenMP] Skip host omp ops when compiling for the target device
This patch separates the lowering dispatch for host and target devices. For the
target device, if the current operation is not a top-level operation
(omp.target) or is inside a target device code region it will be ignored,
since it belongs to the host code.

This is a copy of PR85239 upstream.