[flang][OpenMP] Change the lowering of max and min reduction operator.
This patch changes the lowering of max and min to be lowered to
arith::MaxNumFop and arith::MinNumFOp instead of using arith::MaximumFOp and
arith::MinimumFOp to match the code generated in clang. The arith::MaximumFOp
and arith::MinimumFOp map to the corresponding intrinsics llvm.maximum.* and
llvm.minimum.* intrinsics which conform to the semantics specified in the draft
of IEEE 754-2019. which is not supported by all hardware. Instead using
arith::MaximumFOp and arith::MinimumFOp will allow code generation for more
targets and match the code generated by clang OpenMP.