ockl: Don't use wave32 ballot builtin
Wave32 and wave64 paths cannot really co-exist in the same function
or callgraph. They need to be treated as a hard ABI incompatibility.
We cannot handle the wave32 operation on wave64, but we can and do
handle the wave64 operation on wave32. Given the current linking scheme,
the most expedient fix for this not working is to pretend wave32 does
not exist and just use the wave64 ballot. The optimizer will fold
the 64-bit ballot intrinsic to the 32-bit one when it sees a 32-bit
target.

This was reported broken in
https://github.com/llvm/llvm-project/issues/89332

Change-Id: If897f631066e257e18d8adf574cd17d3f9278ca7