[Flang][OpenMP] Remove unnecessary handling for loop wrappers
The code removed by this patch was added with the intention of preventing the
use of regions belonging to loop wrapper operations as a point to insert
allocations. However, this issue is only present for `omp.parallel` wrappers,
which are handled right before this piece of code.

The result is that any operations that are currently captured by this piece of
code would be handled correctly by letting execution of the `getAllocaBlock()`
method continue and return the entry block of the parent `func.func` operation.