[MLIR][OpenMP] Correctly handle branching within target captures
This patch improves the detection of captured OpenMP constructs inside of an
`omp.target` operation by also considering potential branches. If a nested
OpenMP construct can be executed in a loop or optionally omitted by means of
explicit MLIR control flow, then it's not supposed to be captured.

The following Fortran example results in such a case:
```f90
!$omp target teams
do i = 1, n
  !$omp distribute parallel do
  do j = 1, n
    ...
  end do
  !$omp end distribute parallel do
end do
!$omp end target teams
```

The result of lowering that code to MLIR is the creation of multiple blocks and
branches inside of the `omp.teams` operation's region. Without this change, it
is identified as an SPMD kernel during translation to LLVM IR due to the
nesting of operations, but it is a generic kernel, so it causes a compiler
crash. This is because it tries to get host-evaluated loop bounds that do not
exist to calculate the inner loop's trip count.