[OMPIRBuilder] Init default kernel max-threads with -1
This patch removes the initialization of the MaxThreads global value and
related kernel attributes based on the default workgroup size of the target
architecture whenever an explicit value is not specified by the user through
`thread_limit` or `num_threads` clauses.

Instead, leave these undefined cases represented as a -1 value.