[OMPIRBuilder] Improve 'target if' implementation
This patch cleans up support for the 'if' clause on 'target' directives by
reusing the existing `emitIfClause()` function rather than duplicating code.
One side effect of this change is that constant 'if' values will no longer
result in the generation of branches. Instead, code is generated only for the
applicable case.

It also adds a unit test to ensure it works as expected.