[Flang] Disable Global Filtering pass by default
Global Filtering Passs was created to fix linking
issues for old clang driver. This pass is not upstreamed
and we can switch it off for flang-new and the new clang
driver. The user can still turn it by setting
`-foffload-global-filtering' flag.

This is the first step to completely remove global filtering
pass from ATD branch and to minimize delta between ATD and upstream
branches.